day 18 pt 2 :-) (Int/Long!!!!)

+++ src/main/kotlin/y2023/day10/Network.kt
@@ -19,7 +19,7 @@ data class Network(val matrix: List<List<PipeSegment>>) {
     }
 
     private fun nodeAt(location: Point): Node? =
-        matrix.getOrNull(location.y)?.getOrNull(location.x)?.let { p ->
+        matrix.getOrNull(location.y.toInt())?.getOrNull(location.x.toInt())?.let { p ->
             nodes.computeIfAbsent(location) { Node(it, p) }
         }
